+++ src/com/mal/tests/Knapsack.java
@@ -10,7 +10,10 @@ import com.mal.framework.utils.Problem;
 import java.util.HashMap;
 
 public class Knapsack extends Problem {
+
+    //final static double[] leftHandSide = new double[] {70,73,77,80,82,87,90,94,98,106,110,113,115,118,120};
     final static double[] leftHandSide = new double[] {120,118,115,113,110,106,98,94,90,87,82,80,77,73,70};
+    //final static double[] dataset      = new double[] {135,139,149,150,156,163,173,184,192,201,210,214,221,229,240};
     final static double[] dataset      = new double[] {240,229,221,214,210,201,192,184,173,163,156,150,149,139,135};
 
     private Knapsack(Constraint[] constraints, Bound bound, NodeStrategy strategy, ProblemType type, boolean lp) {
@@ -18,9 +21,8 @@ public class Knapsack extends Problem {
     }
 
     private static class knapsackBounds implements Bound {
-
         @Override
-        public float Lowerbound(HashMap<Integer,Float> currentSolution, double[] set, Constraint[] constraints) {
+        public double Lowerbound(HashMap<Integer,Double> currentSolution, double[] set, Constraint[] constraints) {
             float sum = 0;
             float weight = 0;
             for (int cs_i:currentSolution.keySet()){
@@ -39,7 +41,7 @@ public class Knapsack extends Problem {
         }
 
         @Override
-        public float Upperbound(HashMap<Integer,Float> currentSolution, double[] set, Constraint[] constraints) {
+        public double Upperbound(HashMap<Integer,Double> currentSolution, double[] set, Constraint[] constraints) {
             return 0;
         }
     }
